requirements.txt
fastapi
uvicorn
pydantic
azure-storage-blob
pandas
pyarrow
typing
duckdb
dotenv